pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asgml"
version = "0.1.0"
description = "Asynchronous Staleness Guided Multimodal Learning"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Anonymous", email = "anonymous@email.com"}
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "librosa>=0.9.0",
    "opencv-python>=4.5.0",
    "pillow>=8.0.0",
    "tqdm>=4.62.0",
    "tensorboard>=2.10.0",
    "wandb>=0.13.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[project.scripts]
asgml-train = "scripts.train:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "configs*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
